backend/requirements.txt
asgiref==3.5.0
Django==3.2.16
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
flake8==5.0.4
fpdf==1.7.2
gunicorn==20.0.4
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.1.0
python-dotenv==0.21.0
pytz==2020.1
sqlparse==0.3.1